[SNAPPYDATA] fixing all failures in snappy-spark test suite
Three broad categories of issues fixed:

- handling of double values in JSON conversion layer of the metrics; upstream spark has all
  metrics as Long but snappy-spark has the timings one as double to give more accurate results
- library version differences between Spark's maven poms and SnappyData's gradle builds;
  these are as such not product issues but this checkin changes some versions to be
  matching to maven builds to be fully upstream compatible
- path differences in test resource files/jars when run using gradle rather than using maven

Other fixes and changes:

- the optimized Decimal.equals gave incorrect result in case the scale of the two is different;
  this followed the Java BigDecimal convention of returning false if the scale is different
  but that is incorrect as per Spark's conventions; this should normally not happen from catalyst
  layer but can happen in RDD operations
- correct accumulator result in Task to be empty rather than null when nothing present
- override the extended two argument DStream.initialize in MapWithStateDStream.initialize
- correct the UI path for Spark cache to be "/Spark Cache/" rather than "/storage/"
- avoid sending the whole child plan across in DeserializeToObjectExec to executors when
  only the output is required (also see SNAP-1840 caused due to this)
- rounding of some of the time statistics (that are accumulated as double) in Spark metrics
- SparkListenerSuite local metrics tests frequently failed due to deserialization time being zero
  (despite above change); the reason being the optimizations in snappy-spark that allow it to
  run much quicker and not registering even with System.nanoTime(); now extended the closure
  to force a 1 milliseond sleep in its readExternal method
- use spark.serializer consistently for data only and spark.closureSerializer for others
  (for the case the two are different)
- don't allow inline message size to exceed spark.rpc.message.maxSize
- revert default spark.locality.wait to be 3s in Spark (will be set at snappydata layer if required)
- make SparkEnv.taskLogger to be serializable if required (extend Spark's Logging trait)
- account for task decompression time in the deserialization time too

The full spark test suite can be run either by:

- ./dev/snappy-build.sh && ./dev/snappy-build.sh test (or equivalent)
- ./gradlew check
- from SnappyData:
  - ./gradlew snappy-spark:check, OR
  - ./gradlew precheckin -Pspark (for full test suite run including snappydata suite)

For SnappyData product builds, one of the last two ways from SnappyData should be used

+++ core/src/main/scala/org/apache/spark/storage/ShuffleBlockFetcherIterator.scala
@@ -316,7 +316,8 @@ final class ShuffleBlockFetcherIterator(
     currentResult = results.take()
     val result = currentResult
     val stopFetchWait = System.nanoTime()
-    shuffleMetrics.incFetchWaitTime(math.max(stopFetchWait - startFetchWait, 0L) / 1000000.0)
+    shuffleMetrics.incFetchWaitTime(
+      math.max(stopFetchWait - startFetchWait, 0L).toDouble / 1000000.0)
 
     result match {
       case SuccessFetchResult(_, address, size, buf, isNetworkReqDone) =>
